- process

+++ core/src/main/java/lv/degra/accounting/core/truck/model/TruckRepository.java
@@ -2,5 +2,5 @@ package lv.degra.accounting.core.truck.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TruckRepository extends JpaRepository<Truck, Long> {
+public interface TruckRepository extends JpaRepository<Truck, Integer> {
 }
